refactor: keyboard open function move to utile
